Refactor code somewhat
So we are using a linter now. As a result, the code has changed somewhat
to make sure the linter says everything is okay :). I also removed some
do-notations where the shorter line is equally or more legible, i.e. by
using the (<$>) operator.

Also, rename some things to make sure we use consistent spelling.
Finally, remove some warnings from `cabal build` relating to the fact
that not all modules have their name listed in the executable's
other-modules field. However, I personally think it's kind of strange
that it wants that, given the fact that the executable depends on the
library that does carry those files...