Add parser for TSPLIB files
The main function now refers to a file that is not in the repo; it works
for the "XQF131" problem, the first problem on this page:
http://www.math.uwaterloo.ca/tsp/vlsi/index.html
I haven't tested it on any other files so far. The file is not committed
now because I don't know the license of the data. Better to be safe than
sorry, right? :)

If any other files give problems with the current parser, just tell me
and I'll fix it.