Rename some genetic-related things
The names made no sense anymore, so I made them make sense. Also, the
new Mutation class takes a parameter for the probability of mutation.

A lot of functions now take a probability as a parameter, in order to
make sure the probability of mutation is passed around correctly.

Finally, runGenetic is fully parametrized now.