Add documentation, remove dead code
Documentation added to lots of stuff that I made.

The dead code removal is from the CNF parser, since we had a different
format at first but then decided not to make our own problems anymore.
As a result, we created a new parser, but never removed the old one,
until now.